documentation, transition to i64 coords, start of chain moving

+++ src/block_view.rs
@@ -12,7 +12,7 @@ pub trait BlockView {
 
 pub trait BlockCodeView {
     fn area_size(&self, id: usize) -> (usize, usize);
-    fn block_at(&self, id: usize, x: usize, y: usize) -> Option<&dyn BlockView>;
-    fn origin_at(&self, id: usize, x: usize, y: usize) -> Option<(usize, usize)>;
+    fn block_at(&self, id: usize, x: i64, y: i64) -> Option<&dyn BlockView>;
+    fn origin_at(&self, id: usize, x: i64, y: i64) -> Option<(i64, i64)>;
 }
